requirements.txt
academictorrents==2.3.3
albumentations==1.4.7
einops==0.8.0
imageio==2.34.1
more_itertools==10.2.0
numpy==1.26.4
omegaconf==2.3.0
Pillow==10.3.0
pudb==2024.1
pytorch_lightning==2.2.4
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.13.0
setuptools==69.5.1
skimage==0.0
streamlit==1.34.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
omegaconf==2.3.0